DONE - It works as far as I know

+++ files/php/cronjobs/syncOffers.php
@@ -4,6 +4,7 @@
 # Plus this variable: '/lockfiles/syncOffers.lock' do it the right way so it can be used on every os
 $lockFile = __DIR__. DIRECTORY_SEPARATOR. 'lockfiles'. DIRECTORY_SEPARATOR. 'syncOffers.lock';
 # Variables
+$boolLockFileSystemEnabled = false;
 $accessToken = ['accessToken', '5375636B4D79416363657373546F6B656E'];
 $overrideToken = ['overrideToken', '5375636B4D794C6F636B46696C65546F6B656E'];
 
@@ -22,25 +23,27 @@ if (!$accessAllowed) {
 	die("Access not allowed.");
 }
 
-# Check if this cronjob currently is activated
-if (file_exists($lockFile)) {
-	# Check if an override is activated or not
-	if (!$overrideActivated) {
-		die("The cronjob currently already is running.");
+if ($boolLockFileSystemEnabled) {
+	# Check if this cronjob currently is activated
+	if (file_exists($lockFile)) {
+		# Check if an override is activated or not
+		if (!$overrideActivated) {
+			die("The cronjob currently already is running.");
+		}
 	}
-}
-else {
-	# Check if the directory exists
-	if (!file_exists(dirname($lockFile))) {
-		mkdir(dirname($lockFile), 0777, true);
+	else {
+		# Check if the directory exists
+		if (!file_exists(dirname($lockFile))) {
+			mkdir(dirname($lockFile), 0777, true);
+		}
 	}
+	touch($lockFile);
 }
-touch($lockFile);
 
 // ============ Imports ============
 # WordPress
-if ( $wpLoad = dirname( __DIR__, 6 ) . '/wp/wp-load.php' and file_exists( $wpLoad ) ) {require_once( $wpLoad );}
-elseif ( $wpLoad = dirname( __DIR__, 6 ) . '/wp-load.php' and file_exists( $wpLoad ) ) {require_once( $wpLoad );}
+if ($wpLoad = dirname( __DIR__, 6 ) . '/wp/wp-load.php' and file_exists( $wpLoad ) ) {require_once( $wpLoad );}
+elseif ($wpLoad = dirname( __DIR__, 6 ) . '/wp-load.php' and file_exists( $wpLoad ) ) {require_once( $wpLoad );}
 
 # Classes / Functions
 require_once( dirname( __DIR__, 1 ) . '/includes/classes.php' );
@@ -59,5 +62,7 @@ global $wpdb;
 $ogOffers = new OGOffers();
 
 // ============ End of Program ============
-unlink($lockFile);
+if ($boolLockFileSystemEnabled) {
+	unlink($lockFile);
+}
 ?>
